+++ components/src/lib/index.ts
@@ -1,3 +1,4 @@
 export { default as Modal } from './components/Modal';
 export { default as useModal } from './hooks/useModal';
+export { default as useModalCloseByEsc } from './hooks/useModalCloseByEsc';
 export * from './types/Modal.type';
